adds in order traversal of the tree
meaning you can traverse the tree
and get the ordered values as
a result of the insertion into this
Binary Search Tree (BST).

Also adds pre/post ordered travesal.
Pre (I think) is the order at which
the elements were added to the BST.

Not sure how/when to use the post
ordered traversal - yet.